Intro to Swift
Basic Assignment to gain fluency in Swift. Create a program in Swift to decide where to eat out based on your dining budget. This budget planner will run for the 14 weeks of this semester. You have a weekly budget of $50 plus any left over dining money from the previous week that rolls over. For every day in the week, you will 1) choose a random restaurant from a predetermined list, 2) calculate tax and tip for the given cost of a meal, and 3) determine if eating there is within your budget. If it is not, you will choose a different restuarant and see if that one meets your budget constraints. If it is, you will take the cost of dining there out of your current budget, and print the following line to the Playground's console: 
"[day of the week] meal: [restaurant], budget now [current budget]"